% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_surv_box_plot.R
\name{box_surv}
\alias{box_surv}
\title{box_surv}
\usage{
box_surv(exp_hub, exprSet_hub, meta)
}
\arguments{
\item{exp_hub}{an expression matrix for hubgenes}

\item{exprSet_hub}{a tumor expression set for hubgenes}

\item{meta}{meta data corresponds to expression set}
}
\value{
patchwork result for hub genes boxplot and survival plot
}
\description{
draw box plot for a hub gene expression matrix
}
\examples{
k = box_surv(log2(exp_hub1+1),exprSet_hub1,meta1);k[[1]]
}
\seealso{
\code{\link{exp_boxplot}};\code{\link{exp_surv}}
}
\author{
Xiaojie Sun
}
