% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod}
\alias{cod}
\title{cod}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 512 columns.
}
\usage{
cod
}
\description{
An expression matrix form TCGA
}
\examples{
  cod
}
\keyword{datasets}
