% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.full.R
\name{cor.full}
\alias{cor.full}
\title{cor.test for all varibles}
\usage{
cor.full(x)
}
\arguments{
\item{x}{A numeric matrix or data.frame}
}
\value{
a data.frame with cor.test p.value and estimate
}
\description{
cor.test for all varibles
}
\examples{
x = iris[,-5]
cor.full(x)
}
\seealso{
\code{\link{draw_heatmap}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
