% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.one}
\alias{cor.one}
\title{cor.test for one varible with all varibles}
\usage{
cor.one(x, var)
}
\arguments{
\item{x}{A numeric matrix or data.frame}

\item{var}{your choosed varible,only one.}
}
\value{
A data.frame with cor.test p.value and estimate
}
\description{
cor.test for all varibles(each two columns)
}
\examples{
x = iris[,-5]
cor.one(x,"Sepal.Width")
}
\seealso{
\code{\link{draw_heatmap}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
