% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_plots.R
\name{corheatmap}
\alias{corheatmap}
\title{corheatmap}
\usage{
corheatmap(exp, x, y, color = c("#2fa1dd", "white", "#f87669"))
}
\arguments{
\item{exp}{A numeric matrix}

\item{x}{genes or cells from exp}

\item{y}{genes or cells from exp}

\item{color}{color for heatmap}
}
\value{
a ggplot object
}
\description{
draw cor heatmap
}
\examples{
x = rownames(exprSet_hub1)[1:3]
y = rownames(exprSet_hub1)[4:7]
corheatmap(exprSet_hub1,x,y)
}
\author{
Xiaojie Sun
}
