% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg}
\alias{deg}
\title{deg}
\format{
An object of class \code{data.frame} with 18591 rows and 10 columns.
}
\usage{
deg
}
\description{
limma differential analysis result for GSE42872
}
\examples{
  head(deg)
}
\keyword{datasets}
