% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deseq_data}
\alias{deseq_data}
\title{deseq_data}
\format{
An object of class \code{data.frame} with 552 rows and 6 columns.
}
\usage{
deseq_data
}
\description{
DEseq2 differential analysis result
}
\examples{
  head(deseq_data)
}
\keyword{datasets}
