% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_quick_double_enrich.R
\name{double_enrich}
\alias{double_enrich}
\title{draw enrichment bar plots for both up and down genes}
\usage{
double_enrich(deg, n = 10, color = c("#2874C5", "#f87669"))
}
\arguments{
\item{deg}{a data.frame contains at least two columns:"ENTREZID" and "change"}

\item{n}{how many terms will you perform for up and down genes respectively}

\item{color}{color for bar plot}
}
\value{
a list with kegg and go bar plot according to up and down genes enrichment result.
}
\description{
draw enrichment bar plots for both up and down genes,for human only.
}
\examples{
\dontrun{
if(requireNamespace("org.Hs.eg.db",quietly = TRUE)&
   requireNamespace("labeling",quietly = TRUE)){
   double_enrich(deg)
}else{
  if(!requireNamespace("org.Hs.eg.db",quietly = TRUE)) {
    warning("Package 'org.Hs.eg.db' needed for this function to work.
        Please install it by BiocManager::install('org.Hs.eg.db')",call. = FALSE)
  }
  if(!requireNamespace("labeling",quietly = TRUE)) {
    warning("Package 'labeling' needed for this function to work.
        Please install it by install.packages('labeling')",call. = FALSE)
  }
}
}
}
\seealso{
\code{\link{quick_enrich}}
}
\author{
Xiaojie Sun
}
