% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_deg.R
\name{draw_heatmap2}
\alias{draw_heatmap2}
\title{draw heatmap plots}
\usage{
draw_heatmap2(
  exp,
  group_list,
  deg,
  heat_union = T,
  heat_id = 1,
  gene_number = 200,
  scale_before = F,
  n_cutoff = 2,
  cluster_cols = T,
  annotation_legend = F
)
}
\arguments{
\item{exp}{A numeric matrix}

\item{group_list}{A factor with duplicated character or factor}

\item{deg}{a data.frame created by Differential analysis}

\item{heat_union}{logical , if TRUE ,use union or intersect DEGs for heatmap}

\item{heat_id}{id of heatmap,1 for all DEGs,2 for head and tail,3 for top n DEGs}

\item{gene_number}{how many DEGs will heatmap show .}

\item{scale_before}{logical,if TRUE ,scale before plot,if FALSE,ues scale = "row" param form pheatmap}

\item{n_cutoff}{2 by defalut , scale before plot and set a cutoff,usually 2 or 1.6}

\item{cluster_cols}{if F,heatmap will nor cluster in column}

\item{annotation_legend}{logical，show annotation legend or not}
}
\value{
a heatmap plot according to \code{exp} and grouped by \code{group}.
}
\description{
print heatmap plots for expression matrix and group by group_list paramter
}
\examples{
gse = "GSE474"
geo = geo_download(gse)
geo$exp[1:4,1:4]
geo$exp=log2(geo$exp+1)
library(stringr)
group_list=ifelse(str_detect(geo$pd$title,"MObese"),"MObese",ifelse(str_detect(geo$pd$title,"NonObese"),"NonObese","Obese"))
group_list=factor(group_list,levels = c("NonObese","Obese","MObese"))
find_anno(geo$gpl)
ids <- toTable(hgu133aSYMBOL)
deg = multi_deg(geo$exp,group_list,ids,adjust = F)
draw_heatmap2(geo$exp,group_list,deg)
}
\seealso{
\code{\link{draw_pca}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
