% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{draw_volcano}
\alias{draw_volcano}
\title{draw a volcano plot}
\usage{
draw_volcano(
  deg,
  lab = NA,
  pvalue_cutoff = 0.05,
  logFC_cutoff = 1,
  pkg = 1,
  adjust = F,
  symmetry = F
)
}
\arguments{
\item{deg}{a data.frame created by Differential analysis}

\item{pvalue_cutoff}{Cutoff value of pvalue,0.05 by defult.}

\item{logFC_cutoff}{Cutoff value of logFC,1 by defult.}

\item{pkg}{a integer ,means which Differential analysis packages you used,we support three packages by now, 1,2,3,4 respectively means "DESeq2","edgeR","limma(voom)","limma"}

\item{adjust}{a logical value, would you like to use adjusted pvalue to draw this plot,FAlSE by defult.}

\item{symmetry}{a logical value ,would you like to get your plot symmetrical}
}
\value{
a volcano plot according to logFC and P.value(or adjust P.value)
}
\description{
print a volcano plot for Differential analysis result in data.frame fomat.
}
\examples{
data("des")
head(deseq_data)
draw_volcano(deseq_data)
draw_volcano(deseq_data,pvalue_cutoff = 0.01,logFC_cutoff = 2)
}
\seealso{
\code{\link{draw_heatmap}};\code{\link{draw_pca}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
