% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df_dumd.R
\name{dumd}
\alias{dumd}
\title{count unique values in every colunms for data.frame}
\usage{
dumd(x)
}
\arguments{
\item{x}{A data.frame.}
}
\value{
The simple data.frame of columns unique values count in \code{x}
}
\description{
in geo analysis,this function can help you simplify pdata, delete columns with unique values,which can't be used as group vector
}
\section{just }{

See what are you doing
}

\examples{
dumd(iris)
data(ToothGrowth)
x = ToothGrowth
dumd(ToothGrowth)
}
