% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df_dumd.R
\name{edges_to_nodes}
\alias{edges_to_nodes}
\title{edges_to_nodes}
\usage{
edges_to_nodes(edges)
}
\arguments{
\item{edges}{data.frame}
}
\value{
nodes data.frame
}
\description{
get nodes from edges
}
\examples{
df = data.frame(a = c("gene1","gene2","gene3"),
b = c("d,f,a,b",
"c,e,g",
"a,b,d"))
edges = interaction_to_edges(df)
nodes = edges_to_nodes(edges)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
