% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exists_anno_list}
\alias{exists_anno_list}
\title{exists_anno_list}
\format{
An object of class \code{character} of length 175.
}
\usage{
exists_anno_list
}
\description{
AnnoProbe supported  GPLs
}
\examples{
  exists_anno_list
}
\keyword{datasets}
