% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_hub1}
\alias{exp_hub1}
\title{exp_hub1}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 350 columns.
}
\usage{
exp_hub1
}
\description{
An expression matrix from TCGA and Gtex
}
\examples{
  exp_hub1[1:4,1:4]
}
\keyword{datasets}
