% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprSet_hub1}
\alias{exprSet_hub1}
\title{exprSet_hub1}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 177 columns.
}
\usage{
exprSet_hub1
}
\description{
An cpm expression matrix from TCGA,tumor samples only
}
\examples{
  exprSet_hub1[1:4,1:4]
}
\keyword{datasets}
