% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deg_all.R
\name{find_anno}
\alias{find_anno}
\title{find annotation package or files}
\usage{
find_anno(gpl)
}
\arguments{
\item{gpl}{a gpl accession}
}
\value{
a list with deg data.frame, volcano plot and a list with DEGs.
}
\description{
find gpl annotation package or files
}
\examples{
find_anno(GPL570)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
