% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_geo_id.R, R/tmp.R
\name{geo_parser}
\alias{geo_parser}
\title{geo_parser}
\usage{
geo_parser(gse, destdir = getwd())

geo_parser(gse, destdir = getwd())
}
\arguments{
\item{gse}{gse assession number}

\item{destdir}{The destination directory for data downloads}
}
\value{
an ExpressionSet object list

an ExpressionSet object list
}
\description{
download gse data and get informations

download gse data and get informations
}
\examples{
\dontrun{
if(requireNamespace("GEOquery",quietly = TRUE)){
  gse = "GSE42872"
  a = geo_download(gse,destdir=tempdir())
}else{
  print("Package 'GEOquery' needed for this function to work.
         Please install it by BiocManager::install('GEOquery')")
}
}
\dontrun{
if(requireNamespace("GEOquery",quietly = TRUE)){
  gse = "GSE42872"
  a = geo_download(gse,destdir=tempdir())
}else{
  print("Package 'GEOquery' needed for this function to work.
         Please install it by BiocManager::install('GEOquery')")
}
}
}
\seealso{
\code{\link{find_anno}}

\code{\link{find_anno}}
}
\author{
Xiaojie Sun
}
