% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_split_deg.R
\name{get_cgs}
\alias{get_cgs}
\title{get_cgs}
\usage{
get_cgs(deg)
}
\arguments{
\item{deg}{a data.frame created by Differential analysis}
}
\value{
a list with upgenes,downgenes,diffgenes.
}
\description{
extract DEGs from deg data.frame
}
\examples{
#two group
gse = "GSE42872"
geo = geo_download(gse)
group_list = rep(c("A","B"),each = 3)
ids = AnnoProbe::idmap('GPL6244')
deg = get_deg(geo$exp,group_list,ids)
cgs = get_cgs(deg)
#mutigroup
gse = "GSE474"
geo = geo_download(gse)
geo$exp[1:4,1:4]
geo$exp=log2(geo$exp+1)
library(stringr)
group_list=ifelse(str_detect(geo$pd$title,"MObese"),"MObese",ifelse(str_detect(geo$pd$title,"NonObese"),"NonObese","Obese"))
group_list=factor(group_list,levels = c("NonObese","Obese","MObese"))
find_anno(geo$gpl,install = T)
ids <- toTable(hgu133aSYMBOL)
deg = multi_deg(geo$exp,group_list,ids,adjust = F)
cgs = get_cgs(deg)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
