% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_geo_id.R
\name{get_count_txt}
\alias{get_count_txt}
\title{get count from GEO}
\usage{
get_count_txt(gse, destdir = getwd(), download = FALSE)
}
\arguments{
\item{gse}{gse assession number}

\item{destdir}{The destination directory for data downloads.}

\item{download}{download the txt file or not}
}
\value{
a list with deg data.frame, volcano plot and a list with DEGs.
}
\description{
get RNA-seq count file from GEO database
}
\examples{
get_count_txt("GSE162550",destdir = tempdir())
}
\seealso{
\code{\link{geo_download}}
}
\author{
Xiaojie Sun
}
