% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_geo_id.R
\name{get_gpl_txt}
\alias{get_gpl_txt}
\title{get gpl txt from GEO}
\usage{
get_gpl_txt(gpl, destdir = getwd(), download = F)
}
\arguments{
\item{destdir}{The destination directory for data downloads.}
}
\value{
a list with deg data.frame, volcano plot and a list with DEGs.
}
\description{
get gpl annotation txt file from GEO database
}
\examples{
get_gpl_txt("GSE162550",destdir = tempdir())
}
\seealso{
\code{\link{geo_download}}
}
\author{
Xiaojie Sun
}
