% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_network_edge_node.R
\name{interaction_to_edges}
\alias{interaction_to_edges}
\title{interaction_to_edges}
\usage{
interaction_to_edges(df, a = 1, b = 2, sep = ",")
}
\arguments{
\item{df}{interactions data.frame}

\item{a}{column to replicate}

\item{b}{column to split}

\item{sep}{a character string to separate b column}
}
\value{
a new df with two column ,one interaction by one rows
}
\description{
split interactions by sep paramter,return edges data.frame
}
\examples{
df = data.frame(a = c("gene1","gene2","gene3"),
b = c("d,f,a,b",
"c,e,g",
"a,b,d"))
interaction_to_edges(df)
}
\seealso{
\code{\link{edges_to_nodes}}
}
\author{
Xiaojie Sun
}
