% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_dumd_all.R
\name{intersect_all}
\alias{intersect_all}
\title{intersect_all}
\usage{
intersect_all(...)
}
\arguments{
\item{...}{some vectors or a list with some vectors}
}
\value{
vector
}
\description{
calculate intersect  set for two or more elements
}
\examples{
x1 = letters[1:4]
x2 = letters[3:6]
x3 = letters[3:4]
re =intersect_all(x1,x2,x3)
re2 = intersect_all(list(x1,x2,x3))
re3 = union_all(x1,x2,x3)
}
\seealso{
\code{\link{union_all}}
}
\author{
Xiaojie Sun
}
