% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lnc_anno}
\alias{lnc_anno}
\title{lnc_anno}
\format{
An object of class \code{data.frame} with 14826 rows and 3 columns.
}
\usage{
lnc_anno
}
\description{
annotation for TCGA expression matrix(lncRNA),form genecode v22 gtf file.
}
\examples{
  head(lnc_anno)
}
\keyword{datasets}
