% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lnc_annov23}
\alias{lnc_annov23}
\title{lnc_annov23}
\format{
An object of class \code{data.frame} with 14852 rows and 3 columns.
}
\usage{
lnc_annov23
}
\description{
annotation for TCGA and gtex expression matrix(lncRNA),form genecode v23 gtf file.
}
\examples{
  head(lnc_annov23)
}
\keyword{datasets}
