% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tcga_group_trans_exp.R
\name{make_tcga_group}
\alias{make_tcga_group}
\title{make_tcga_group}
\usage{
make_tcga_group(exp)
}
\arguments{
\item{exp}{tcga or tcga_gtex expression set from gdc or xena}
}
\value{
a group factor with normal and tumor ,conresponse to colnames for expression matrtix
}
\description{
make tcga group for given tcga expression matrix
}
\examples{
k = make_tcga_group(exp_hub1);table(k)
}
\seealso{
\code{\link{sam_filter}};\code{\link{match_exp_cl}}
}
\author{
Xiaojie Sun
}
