% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_exp.R
\name{match_exp_cl}
\alias{match_exp_cl}
\title{match_exp_cl}
\usage{
match_exp_cl(exp, cl, id_column = "id", sample_centric = T)
}
\arguments{
\item{exp}{tcga  expression set}

\item{cl}{tcga clinical data.frame}

\item{id_column}{which column containes patient ids, column number or colnmn name.}

\item{sample_centric}{logical,deault T,keep all samples from the same patients.if FALSE,keep only one tumor sample for one patient.}
}
\value{
a transformed clinical data.frame with sample ids.
}
\description{
match exp and clinical data from tcga
}
\examples{
match_exp_cl(exp_hub1,meta1[,2:4],"X_PATIENT")
match_exp_cl(exp_hub1,meta1[,2:4],"X_PATIENT",sample_centric = F)
}
\seealso{
\code{\link{make_tcga_group}};\code{\link{sam_filter}};\code{\link{trans_array}}
}
\author{
Xiaojie Sun
}
