% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta1}
\alias{meta1}
\title{meta1}
\format{
An object of class \code{data.frame} with 177 rows and 4 columns.
}
\usage{
meta1
}
\description{
clinical messages for some TCGA patients,correspond to exprSet_hub1
}
\examples{
  head(meta1)
}
\keyword{datasets}
