% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA_anno}
\alias{mRNA_anno}
\title{mRNA_anno}
\format{
An object of class \code{data.frame} with 19814 rows and 3 columns.
}
\usage{
mRNA_anno
}
\description{
annotation for TCGA and gtex expression matrix(mRNA),form genecode v22 gtf file.
}
\examples{
  head(mRNA_anno)
}
\keyword{datasets}
