% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA_annov23}
\alias{mRNA_annov23}
\title{mRNA_annov23}
\format{
An object of class \code{data.frame} with 19797 rows and 3 columns.
}
\usage{
mRNA_annov23
}
\description{
annotation for TCGA and gtex expression matrix(mRNA),form genecode v23 gtf file.
}
\examples{
  head(mRNA_annov23)
}
\keyword{datasets}
