% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkg_all}
\alias{pkg_all}
\title{pkg_all}
\format{
An object of class \code{data.frame} with 85 rows and 3 columns.
}
\usage{
pkg_all
}
\description{
bioconductor annotation packages for GPLs
}
\examples{
  head(pkg_all)
}
\keyword{datasets}
