% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_cor_and_hyper.R
\name{plcortest}
\alias{plcortest}
\title{plcortest}
\usage{
plcortest(lnc_exp, mRNA_exp, cor_cutoff = 0)
}
\arguments{
\item{lnc_exp}{lncRNA expression set}

\item{mRNA_exp}{mRNA expression set which nrow equal to lncRNA_exp}

\item{cor_cutoff}{cor estimate cut_off, default 0}
}
\value{
a list with cor.test result,names are lncRNAs, element are mRNAs
}
\description{
make cor.test for given lncRNA and mRNA
}
\examples{
# to update
}
\seealso{
\code{\link{hypertest}}
}
\author{
Xiaojie Sun
}
