% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_enrich.R
\name{quick_enrich}
\alias{quick_enrich}
\title{quick_enrich}
\usage{
quick_enrich(genes, kkgo_file = "kkgo_file.Rdata")
}
\arguments{
\item{genes}{a gene symbol or entrizid vector}

\item{kkgo_file}{Rdata filename for kegg and go result}
}
\value{
enrichment reslut and dotplots
}
\description{
do diffiencial analysis according to exprission set and group information
}
\examples{
gse = "GSE42872"
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
