% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tcga_group_trans_exp.R
\name{sam_filter}
\alias{sam_filter}
\title{sam_filter}
\usage{
sam_filter(exp)
}
\arguments{
\item{exp}{tcga or tcga_gtex expression set from gdc or xena}
}
\value{
a transformed expression set without duplicated samples
}
\description{
drop duplicated samples from the same patients
}
\examples{
cod[1:4,1:4]
dim(cod)
cod2 = sam_filter(cod)
dim(cod2)
g = make_tcga_group(cod);table(g)
library(stringr)
table(!duplicated(str_sub(colnames(cod[,g=="tumor"]),1,12)))
}
\seealso{
\code{\link{make_tcga_group}};\code{\link{match_exp_cl}}
}
\author{
Xiaojie Sun
}
