% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df_dumd.R
\name{split_df}
\alias{split_df}
\title{split_df}
\usage{
split_df(df, a = 1, b = 2, sep = ",")
}
\arguments{
\item{df}{data.frame with at least two column}

\item{a}{column to replicate}

\item{b}{column to split}

\item{sep}{a character string to separate b column}
}
\value{
a new df with two column ,one interaction by one rows
}
\description{
split one column by sep paramter,replicate anothor column to return a two-column df
}
\examples{
df = data.frame(a = c("gene1","gene2","gene3"),
b = c("d,f,a,b",
"c,e,g",
"a,b,d"))
split_df(df)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
