% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_dumd_all_split.R
\name{split_list}
\alias{split_list}
\title{Split the list into sublists}
\usage{
split_list(l)
}
\arguments{
\item{l}{a list with more than one element}
}
\value{
single sublists
}
\description{
Split the list into sublists
}
\examples{
m = list(x = 1:2,
         y = matrix(1:10,2),
         z = iris)
split_list(m)
}
\seealso{
\code{\link{dumd}}
}
\author{
Xiaojie Sun
}
