% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_cut_cox_km.R
\name{surv_cox}
\alias{surv_cox}
\title{surv_cox}
\usage{
surv_cox(
  exprSet_hub,
  meta,
  cut.point = F,
  pvalue_cutoff = 0.05,
  HRkeep = "all"
)
}
\arguments{
\item{HRkeep}{one of "all","protect"or"risk"}
}
\value{
a matrix with gene names ,cox p value and HR
}
\description{
caculate cox p values and HR for genes
}
\examples{
surv_cox(exprSet_hub1,meta1,cut.point = T,HRkeep = "all")
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
