% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_cut_cox_km.R
\name{surv_KM}
\alias{surv_KM}
\title{surv_KM}
\usage{
surv_KM(exprSet_hub, meta, cut.point = F, pvalue_cutoff = 0.05)
}
\arguments{
\item{exprSet_hub}{a tumor expression set for hubgenes}

\item{meta}{meta data corresponds to expression set}

\item{cut.point}{logical , use cut_point or not, if FALSE,use median by defult}

\item{pvalue_cutoff}{p value cut off ,0.05 by defult}
}
\value{
a vector with gene names and log_rank p value
}
\description{
caculate log_rank test p values for genes
}
\examples{
surv_KM(exprSet_hub1,meta1)
surv_KM(exprSet_hub1,meta1,cut.point = T)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
