% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_choose.R
\name{t_choose}
\alias{t_choose}
\title{t_choose}
\usage{
t_choose(genes, exp, group_list, up_only = F, down_only = F)
}
\arguments{
\item{genes}{a vector with some genes}

\item{exp}{A numeric matrix}

\item{group_list}{A factor with duplicated character or factor}

\item{up_only}{keep up genes in the result only}

\item{down_only}{keep down genes in the result only}
}
\value{
a vector with diffrencial expressed genes
}
\description{
choose diffrencial expressed genes by simple t.test
}
\examples{
exp = matrix(rnorm(1000),ncol = 10)
rownames(exp) = sample(mRNA_annov23$gene_id,100)
colnames(exp) = c(paste0("TCGA",1:5),paste0("GTEX",1:5))
k  = trans_exp(exp)
k[,1:5] = k[,1:5]+10
exp = k
group_list = rep(c("A","B"),each = 5)
genes = sample(rownames(k),3)
t_choose(genes,k,group_list)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
