% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_cor_t_choose.R
\name{cor.one}
\alias{cor.one}
\title{cor.test for one variable with all variables}
\usage{
cor.one(x, var)
}
\arguments{
\item{x}{A numeric matrix or data.frame}

\item{var}{your chosen variable,only one.}
}
\value{
A data.frame with cor.test p.value and estimate
}
\description{
cor.test for all variables(each two columns)
}
\examples{
x = iris[,-5]
cor.one(x,"Sepal.Width")
}
\seealso{
\code{\link{cor.full}}
}
\author{
Xiaojie Sun
}
