% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_enrich.R
\name{double_enrich}
\alias{double_enrich}
\title{draw enrichment bar plots for both up and down genes}
\usage{
double_enrich(deg, n = 10)
}
\arguments{
\item{deg}{a data.frame contains at least two columns:"ENTREZID" and "change"}

\item{n}{how many terms will you perform for up and down genes respectively}
}
\value{
a list with kegg and go bar plot according to up and down genes enrichment result.
}
\description{
draw enrichment bar plots for both up and down genes
}
\examples{
double_enrich(deg)
}
\seealso{
\code{\link{draw_heatmap}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
