% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{draw a heatmap plot}
\usage{
draw_heatmap(n, group_list, scale_before = F, n_cutoff = 2, cluster_cols = T)
}
\arguments{
\item{group_list}{A factor with duplicated character or factor}

\item{scale_before}{logical,if TRUE ,scale before plot,if FALSE,ues scale = "row" param form pheatmap}

\item{n_cutoff}{2 by defalut , scale before plot and set a cutoff,usually 2 or 1.6}

\item{cluster_cols}{if F,heatmap will nor cluster in column}
}
\value{
a heatmap plot according to \code{exp} and grouped by \code{group}.
}
\description{
print a heatmap plot for expression matrix and group by group_list paramter
}
\examples{
#use your example data
exp = matrix(abs(rnorm(60,sd = 16)),nrow = 10)
exp[,1:3] <- exp[,1:3]+20
colnames(exp) <- paste0("sample",1:6)
rownames(exp) <- paste0("gene",1:10)
exp[1:4,1:4]
group_list = rep(c("A","B"),each = 3)
draw_heatmap(exp,group_list)
#use iris
n = t(iris[,1:4]);colnames(n) = 1:150
group_list = iris$Species
draw_heatmap(n,group_list)
}
\seealso{
\code{\link{draw_pca}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
