% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{draw_pca}
\alias{draw_pca}
\title{draw PCA plots}
\usage{
draw_pca(exp, group_list)
}
\arguments{
\item{exp}{A numeric matrix}

\item{group_list}{A factor with duplicated character or factor}
}
\value{
a pca plot according to \code{exp} and grouped by \code{group}.
}
\description{
do PCA analysis and print a PCA plot
}
\examples{
draw_pca(t(iris[,1:4]),iris$Species)
exp <-  matrix(rnorm(60),nrow = 10)
colnames(exp) <- paste0("sample",1:6)
rownames(exp) <- paste0("gene",1:10)
exp[1:4,1:4]
group_list <- rep(c("A","B"),each = 3)
draw_pca(exp,group_list)
}
\seealso{
\code{\link{draw_heatmap}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
