% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{draw_venn}
\alias{draw_venn}
\title{draw a venn plot}
\usage{
draw_venn(x, name)
}
\arguments{
\item{x}{a list for plot}

\item{name}{main of the plot}
}
\value{
a venn plot according to \code{x}, \code{y} and.\code{z} named "name" paramter
}
\description{
print a venn plot for deg result created by three packages
}
\examples{
x = list(Deseq2=sample(1:100,30),edgeR = sample(1:100,30),limma = sample(1:100,30))
draw_venn(x,"test")
}
\seealso{
\code{\link{draw_pca}};\code{\link{draw_volcano}};\code{\link{draw_heatmap}}
}
\author{
Xiaojie Sun
}
