% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_surv_box_plot.R
\name{exp_surv}
\alias{exp_surv}
\title{exp_surv}
\usage{
exp_surv(exprSet_hub, meta, color = c("grey", "red"))
}
\arguments{
\item{exprSet_hub}{a tumor expression set for hubgenes}

\item{meta}{meta data corresponds to expression set}

\item{color}{color for boxplot}
}
\value{
survival plots list for all genes
}
\description{
draw surv plot for a hub gene expression matrix for tumor samples
}
\examples{
tmp = exp_surv(exprSet_hub1,meta1)
patchwork::wrap_plots(tmp)
}
\seealso{
\code{\link{exp_boxplot}};\code{\link{box_surv}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
