% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_surv_box_plot.R
\name{exp_surv}
\alias{exp_surv}
\title{exp_surv}
\usage{
exp_surv(exprSet_hub, meta, cut.point = FALSE, color = c("#2874C5", "#f87669"))
}
\arguments{
\item{exprSet_hub}{a tumor expression set for hubgenes}

\item{meta}{meta data corresponds to expression set}

\item{cut.point}{logical , use cut_point or not, if FALSE,use median by defult}

\item{color}{color for boxplot}
}
\value{
survival plots list for all genes
}
\description{
draw surv plot for a hub gene expression matrix for tumor samples
}
\examples{
tmp = exp_surv(exprSet_hub1,meta1)
patchwork::wrap_plots(tmp)+patchwork::plot_layout(guides = "collect")
tmp2 = exp_surv(exprSet_hub1,meta1,cut.point = TRUE)
patchwork::wrap_plots(tmp2)+patchwork::plot_layout(guides = "collect")
}
\seealso{
\code{\link{exp_boxplot}};\code{\link{box_surv}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
