% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_geo_id.R
\name{find_anno}
\alias{find_anno}
\title{find annotation package or files}
\usage{
find_anno(gpl, install = FALSE, update = FALSE)
}
\arguments{
\item{gpl}{a gpl accession}

\item{install}{whether to install and library the package}

\item{update}{whether to update the package}
}
\value{
a list with deg data.frame, volcano plot and a list with DEGs.
}
\description{
find gpl annotation package or files
}
\examples{
find_anno("GPL570")
}
\seealso{
\code{\link{geo_download}}
}
\author{
Xiaojie Sun
}
