% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_deg_all.R
\name{find_anno}
\alias{find_anno}
\title{find annotation package or files}
\usage{
find_anno(gpl, install = F, update = F)
}
\arguments{
\item{gpl}{a gpl accession}

\item{install}{whether to install and library the package}

\item{update}{whether to update the package}
}
\value{
a list with deg data.frame, volcano plot and a list with DEGs.
}
\description{
find gpl annotation package or files
}
\examples{
find_anno(GPL570)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
