% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{geo_download}
\alias{geo_download}
\title{geo_download}
\usage{
geo_download(gse, by_annopbrobe = T, simpd = T)
}
\arguments{
\item{gse}{gse assession number}

\item{by_annopbrobe}{getGEO or geoChina}

\item{simpd}{get simplified pdata,drop out columns with all same values}
}
\value{
a list with exp,pd and gpl
}
\description{
download gse data and get informations
}
\examples{
gse = "GSE42872"
geo_download(gse)
geo_download(gse,by_annopbrobe = F)
}
\seealso{
\code{\link{simpd}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
