% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deg.R
\name{get_deg}
\alias{get_deg}
\title{get_deg}
\usage{
get_deg(
  exp,
  group_list,
  ids,
  logFC_cutoff = 1,
  pvalue_cutoff = 0.05,
  adjust = F,
  entriz = T
)
}
\arguments{
\item{exp}{A numeric matrix}

\item{group_list}{A factor with duplicated character or factor}

\item{ids}{a data.frame with 2 columns,including probe_id and symbol}

\item{logFC_cutoff}{Cutoff value of logFC,1 by defult.}

\item{pvalue_cutoff}{Cutoff value of pvalue,0.05 by defult.}

\item{adjust}{a logical value, would you like to use adjusted pvalue to draw this plot,FAlSE by defult.}
}
\value{
a deg data.frame with 10 columns
}
\description{
do diffiencial analysis according to exprission set and group information
}
\examples{
gse = "GSE42872"
geo = geo_download(gse)
Group = rep(c("control","treat"),each = 3)
Group = factor(Group)
find_anno(geo$gpl,install = T)
ids <- toTable(hugene10sttranscriptclusterSYMBOL)
deg = get_deg(geo$exp,Group,ids)
head(deg)
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
