% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_get_deg.R
\name{get_deg}
\alias{get_deg}
\title{get_deg}
\usage{
get_deg(
  exp,
  group_list,
  ids,
  logFC_cutoff = 1,
  pvalue_cutoff = 0.05,
  adjust = FALSE,
  entriz = TRUE
)
}
\arguments{
\item{exp}{A numeric matrix}

\item{group_list}{A factor with duplicated character or factor}

\item{ids}{a data.frame with 2 columns,including probe_id and symbol}

\item{logFC_cutoff}{Cutoff value of logFC,1 by default.}

\item{pvalue_cutoff}{Cutoff value of pvalue,0.05 by default.}

\item{adjust}{a logical value, would you like to use adjusted pvalue to draw this plot,FAlSE by default.}

\item{entriz}{whether convert symbols to entriz ids}
}
\value{
a deg data.frame with 10 columns
}
\description{
do differential analysis according to expression set and group information
}
\examples{
\donttest{gse = "GSE42872"
geo = geo_download(gse,destdir=tempdir())
Group = rep(c("control","treat"),each = 3)
Group = factor(Group)
find_anno(geo$gpl)
ids <- AnnoProbe::idmap(geo$gpl,destdir = tempdir())
deg = get_deg(geo$exp,Group,ids)
head(deg)
}
}
\seealso{
\code{\link{multi_deg}};\code{\link{get_deg_all}}
}
\author{
Xiaojie Sun
}
