% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_and_hyper.R
\name{hypertest}
\alias{hypertest}
\title{hypertest}
\usage{
hypertest(lnc, pc, deMIR = NULL, lnctarget, pctarget)
}
\arguments{
\item{lnc}{lncRNA names}

\item{pc}{mRNA names}

\item{deMIR}{miRNA names , defult NULL}

\item{lnctarget}{a data.frame with two column,lncRNA in the first column ,miRNA in the second column}

\item{pctarget}{a data.frame with two column,mRNA in the first column ,miRNA in the second column}
}
\value{
a data.frame with hypertest result
}
\description{
make hypertest for given lncRNA and mRNA common miRNAs
}
\examples{
to update
}
\seealso{
\code{\link{geo_download}};\code{\link{draw_volcano}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
