% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_multi_deg.R
\name{multi_deg}
\alias{multi_deg}
\title{multi_deg}
\usage{
multi_deg(
  exp,
  group_list,
  ids,
  logFC_cutoff = 1,
  pvalue_cutoff = 0.05,
  adjust = FALSE,
  entriz = T
)
}
\arguments{
\item{exp}{A numeric matrix}

\item{group_list}{A factor with duplicated character or factor}

\item{ids}{a data.frame with 2 columns,including probe_id and symbol}

\item{logFC_cutoff}{Cutoff value of logFC,1 by defult.}

\item{pvalue_cutoff}{Cutoff value of pvalue,0.05 by defult.}

\item{adjust}{a logical value, would you like to use adjusted pvalue to draw this plot,FAlSE by defult.}

\item{entriz}{whether convert symbols to entriz ids}
}
\value{
a deg data.frame with 10 columns
}
\description{
do diffiencial analysis according to exprission set and group information
}
\examples{
gse = "GSE474"
geo = geo_download(gse)
geo$exp[1:4,1:4]
geo$exp=log2(geo$exp+1)
library(stringr)
group_list=ifelse(str_detect(geo$pd$title,"MObese"),
"MObese",ifelse(str_detect(geo$pd$title,"NonObese"),
"NonObese","Obese"))
group_list=factor(group_list,levels = c("NonObese","Obese","MObese"))
find_anno(geo$gpl)
library(hgu133a.db)
ids <- toTable(hgu133aSYMBOL)
deg = multi_deg(geo$exp,group_list,ids,adjust = FALSE)
names(deg)
head(deg[[1]])
head(deg[[2]])
head(deg[[3]])
}
\seealso{
\code{\link{get_deg}};\code{\link{multi_deg_all}}
}
\author{
Xiaojie Sun
}
