% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_quick_double_enrich.R
\name{quick_enrich}
\alias{quick_enrich}
\title{quick_enrich}
\usage{
quick_enrich(genes, kkgo_file = "kkgo_file.Rdata", destdir = getwd())
}
\arguments{
\item{genes}{a gene symbol or entrizid vector}

\item{kkgo_file}{Rdata filename for kegg and go result}

\item{destdir}{destdir to save kkgofile}
}
\value{
enrichment results and dotplots
}
\description{
do diffiencial analysis according to exprission set and group information
}
\examples{
head(genes)
g = quick_enrich(genes,destdir = tempdir())
names(g)
g[[1]][1:4,1:4]
g[[3]]
g[[4]]
}
\seealso{
\code{\link{double_enrich}}
}
\author{
Xiaojie Sun
}
