% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_surv_box_plot.R
\name{risk_plot}
\alias{risk_plot}
\title{risk_plot}
\usage{
risk_plot(
  exprSet_hub,
  meta,
  riskscore,
  cut.point = FALSE,
  color = c("#2fa1dd", "#f87669"),
  n_cutoff = 3
)
}
\arguments{
\item{exprSet_hub}{a tumor expression set for hubgenes}

\item{meta}{meta data corresponds to expression set}

\item{riskscore}{a numeric vector of riskscore}

\item{cut.point}{logical , use cut_point or not, if FALSE,use median by defult}

\item{color}{color for boxplot}

\item{n_cutoff}{3 by defalut , scale before plot and set a cutoff,usually 2 or 1.6}
}
\value{
risk plot
}
\description{
draw risk plot
}
\examples{
risk_plot(exprSet_hub1,meta1,riskscore = rnorm(nrow(meta1)))
}
\seealso{
\code{\link{exp_boxplot}};\code{\link{box_surv}};\code{\link{draw_venn}}
}
\author{
Xiaojie Sun
}
